Category reset

+++ isd-farmers-market/src/main/java/com/example/isdfarmersmarket/dao/repositories/CategoryRepository.java
@@ -1,14 +1,13 @@
 package com.example.isdfarmersmarket.dao.repositories;
 
 import com.example.isdfarmersmarket.dao.models.Category;
-import jakarta.validation.constraints.NotNull;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
 
 @Repository
-public interface CategoryRepository extends JpaRepository<Category, Integer> {
+public interface CategoryRepository extends JpaRepository<Category, Long> {
     Optional<Category> getCategoryById(Long id);
 
     Boolean existsByTitle(String title);
